+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Test of a test task returning several performances ([#251](https://github.com/Substra/substra-tests/pull/251))
+
 ## [0.39.0] - 2023-03-31
 
 ### Changed