+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - Define test client (`substratest.client.Client`) as child class of `substra.Client` (#205)
+([#257](https://github.com/Substra/substra-tests/pull/257))
+
 ## [0.41.0] - 2023-06-12
 
 - Update tests using new substra sdk `list` and `get` functions ([#256](https://github.com/Substra/substra-tests/pull/256))